Clean up the 'check' hook
Previously, the check hook was checking for the cloud-config of a FULL
concourse, even if we had set up our features to deploy a satellite.
For the most part, the `if !defined` logic kept this from being onerous,
but it did still require things like the concourse_disk_type to be found
in cloud-config, even though it was not active.

This also removes some bad `vault_*` cloud-config references, probably
from a copy-pasta of the original check hook.

Fixes #36