Add `small-footprint` feature (#40)
This Genesis Kit can now co-locate all non worker-related jobs to a single VM. It drops HAProxy, but still supports HTTPS through the native web job.

This is a significant rewrite of the kit to accommodate small footprint. To reduce YAML duplication, we moved most jobs to a `jobs.yml` manifest and referenced them in full and small-footprint merges.