Change releases and jobs to inject
Properties could not be overridden at their usual spot in the env file
because there was no name key for spruce to latch onto. Now the name key
is specified for each job, and then the other common configurations are
injected in - which pleases the people using meta and the people
overriding inline. win win win win win win.