Change Cloud-config to use newer form
In the updated OCFP model, the VM names are ocfp_env.vm_machinetype
instead of machinetype-scale.  We updated the code for the AWS path

We added the concourse-web to STACKIT to match AWS but the VM's are
not being moved to that network at this point.  We are looking at
split networks again to see if they provide value on STACKIT